Remove support for string interpolation
The behaviour of the default python ConfigParser changes between python2
and python3. While a value containing a percentage sign ('%') is valid
in python2, it raises a ``configparser.InterpolationSyntaxErrorerror``
in python3:

    '%' must be followed by '%' or '('')'

This commit removes interpolation support with python2 and python3.
As mentioned in the python documentation, with the following file:

    [Paths]
    home_dir: /Users
    my_dir: %(home_dir)s/lumberjack
    my_pictures: %(my_dir)s/Pictures

the parser would simply return %(my_dir)s/Pictures as the value of
my_pictures and %(home_dir)s/lumberjack as the value of my_dir.